% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laurent_approximation.R
\name{laurent_approximation}
\alias{laurent_approximation}
\title{Compute Local Laurent Series Approximation}
\usage{
laurent_approximation(data, center_idx, window_size = 20, n_neg = 3, n_pos = 5)
}
\arguments{
\item{data}{Numeric vector. Time series data}

\item{center_idx}{Integer. Index of the center point}

\item{window_size}{Integer. Number of points on each side of center}

\item{n_neg}{Integer. Number of negative power terms (principal part)}

\item{n_pos}{Integer. Number of positive power terms (regular part)}
}
\value{
A list with the following components:
\itemize{
  \item neg_coeffs: Coefficients of negative powers
  \item const_coeff: Constant term
  \item pos_coeffs: Coefficients of positive powers
  \item principal_norm: Norm of the principal part (singularity indicator)
  \item residue: Residue (coefficient of (t-t0)^{-1})
  \item center_idx: Center index
}
}
\description{
Approximates a time series locally around a center point using Laurent series.
}
\examples{
\dontrun{
data <- rnorm(100)
laurent <- laurent_approximation(data, center_idx = 50)
print(laurent$principal_norm)
}

}
