% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_singularities.R
\name{plot_singularities}
\alias{plot_singularities}
\title{Visualize Detected Singularities}
\usage{
plot_singularities(
  df,
  singularities,
  variable = "LogReturns",
  title = "Financial Singularities Detection"
)
}
\arguments{
\item{df}{Data.frame of financial data}

\item{singularities}{Data.frame of detected singularities}

\item{variable}{Character. Variable to plot}

\item{title}{Character. Plot title}
}
\value{
A ggplot object
}
\description{
Creates a time series plot with singularities highlighted.
}
\examples{
\dontrun{
df <- load_financial_data('AAPL', '2024-01-01', Sys.Date())
sing <- detect_singularities(df)
plot_singularities(df, sing)
}

}
